20 server refactor (#30)
* chore: made new darkmoon directory and placed in files

* chore: created core directory and placed __init__.py inside

* chore: finished setting up basic internal components of new directory

* chore: changes to internal files

* chore: checking changes made to docker

* chore: took out tests for docker

* chore: refactored internal mechanics of darkmoon directory

* refactor: added router, moved functions to views.py

* refactor: checking database / app connection

* refactor: disestablished connection between database and app, removed unnecessary code from app

* chore: created test and unit directories under darkmoon

* refactor: renamed darkmoon directory to server
