+++ server/Dockerfile
@@ -19,11 +19,11 @@ RUN pip install poetry
 COPY pyproject.toml poetry.lock ./
 RUN poetry export --without-hashes -f requirements.txt -o requirements.txt
 RUN pip install -r requirements.txt
-COPY src .
+COPY . .
 RUN poetry build && pip install dist/*.whl
 
 FROM darkmoon-base AS runner
 COPY --from=darkmoon-builder /usr/local/lib/python3.10/site-packages /usr/local/lib/python3.10/site-packages
 ENV PRODUCTION=true \
     STATIC_PATH=/opt/app/
-CMD ["python", "-m", "uvicorn", "darkmoon.main:app", "--host=0.0.0.0", "--port=8000", "--reload"]
+CMD ["python", "-m", "uvicorn", "darkmoon.app:app", "--host=0.0.0.0", "--port=8000", "--reload"]
