+++ server/tests/unit/test_app.py
@@ -1,0 +1,1 @@
+"""This is the testing area for app.py."""
